+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## [Unreleased]
 
-### Changed
+## Changed
 * Target net6.0 for fantomas-tool.
+
+## Fixed
+* SRTP or condition disappear when non-generic type is used. [#2168](https://github.com/fsprojects/fantomas/issues/2168)
+* Multiline AbstractSlot without constraints introduces newline [#2175](https://github.com/fsprojects/fantomas/issues/2175)
 
 ## [5.0.0-alpha-001] - 2022-03-19